requirements.txt
prometheus-client==0.21.0
pyyaml==6.0.2
requests==2.32.3
pytest==8.3.3
flake8==7.1.1
